% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preProcessData.R
\name{preProcessData}
\alias{preProcessData}
\title{preProcessData}
\usage{
preProcessData(file.name = "", impute = FALSE)
}
\arguments{
\item{file.name}{data file name}

\item{impute}{(logical) impute missing values with missforest or drop missing values.
Note imputation with \code{\link[missForest]{missForest}} is going to take a considerable amount of time.}
}
\value{
training data, response variables, predictors and formula.
}
\description{
Preprocess data: transform, normalize, select, impute variables etc.
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

