% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Train.R
\name{Train}
\alias{Train}
\alias{Train.Validate.Test}
\alias{TrainModels}
\title{Train, Validate and Test various models}
\usage{
TrainModels()

Train.Validate.Test(classifier, X.trn, Y.trn, X.val, Y.val, X.tst, Y.tst, para,
  opt.para = FALSE, return.model = FALSE, varimp = TRUE, ...)
}
\arguments{
\item{classifier}{character list of classification models. See names(TrainAllModels()).}

\item{X.trn,X.val,X.tst}{matrix of predictors}

\item{Y.trn,Y.val,Y.tst}{matrix of binary \{0,1\} response variable}

\item{para}{named list of model parameters}

\item{opt.para}{(logical) tune parameters ?}

\item{return.model}{(logical) return trained model ?}

\item{varimp}{(logical) compute variable importance ?}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
\code{TrainModels} returns a list of functions for training various algorithms. Currently
7 machine learning algorithms ELR, GLM, GLMnet, RF, GBM, avNNET, and SVM. Others can be easily added.
\code{Train.Validate.Test} returns a list of performance measures for each classifier traoned:
\item{model}{trained model}
\item{para}{named list of model hyper-paramters (tunned values if opt.para = TRUE)}
\item{run.time}{compute time}
\item{varimp}{variable importance if classifier is GLM, GBM or RF}
\item{perf}{a list with two data.frames: val.perf and tst.perf containing performance measures
for validation and test sets }
}
\description{
Functions to train, select best hyper-paramters, select best classification thresholds
 and evaluate final model on the test set  using a number of machine
learning classification algorithms
}
\examples{
Trainmodels <- lapply(TrainModels(), function(x) x)
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}

